pyproject.toml
[project]
name = "aibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.21.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
llm = [
    "google-generativeai>=0.3.0",
]

[tool.setuptools]
packages = ["backend"]
